refactor: Modularize game architecture with separate Durable Objects per game type
- Extract base GameSession class with common functionality (chat, spectators, player management)
- Create CheckboxGameSession for checkbox-specific game logic
- Create EverybodyVotesGameSession for Everybody Votes MVP implementation
- Extract GameSessionRegistry to separate file
- Update routing to use appropriate Durable Object based on game type
- Clean up index.ts from 1100+ lines to ~200 lines
- Each game now has proper separation of concerns

This sets up a clean, extensible architecture where each game type has its own Durable Object implementation while sharing common functionality through the base GameSession class.
